Split Database access stuff to new project
