+++ DiscordBot.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 16.0.31129.286
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DiscordBot", "DiscordBot\DiscordBot.csproj", "{7D4202AF-1F1D-4EEF-96AD-F5E752D3116E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiscordBotApi", "DiscordBotApi\DiscordBotApi.csproj", "{524F1FC0-79E1-49B1-ACDE-FF9510AAFEFF}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DiscordBotApi", "DiscordBotApi\DiscordBotApi.csproj", "{524F1FC0-79E1-49B1-ACDE-FF9510AAFEFF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiscordBotApiLib", "DiscordBotApiLib\DiscordBotApiLib.csproj", "{7A5101B7-D87C-4F80-AA01-E576FF5531B6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{524F1FC0-79E1-49B1-ACDE-FF9510AAFEFF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{524F1FC0-79E1-49B1-ACDE-FF9510AAFEFF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{524F1FC0-79E1-49B1-ACDE-FF9510AAFEFF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7A5101B7-D87C-4F80-AA01-E576FF5531B6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7A5101B7-D87C-4F80-AA01-E576FF5531B6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7A5101B7-D87C-4F80-AA01-E576FF5531B6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7A5101B7-D87C-4F80-AA01-E576FF5531B6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
