+++ DiscordBotApi/DiscordBotApi.csproj
@@ -19,4 +19,8 @@
     <Folder Include="Migrations\" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\DiscordBotApiLib\DiscordBotApiLib.csproj" />
+  </ItemGroup>
+
 </Project>
