+++ DiscordBotApi/Dtos/ServerLogItemCreateDto.cs
@@ -23,7 +23,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-using DiscordBotApi.Models;
+using DiscordBotApiLib.Models;
 using System.ComponentModel.DataAnnotations;
 
 namespace DiscordBotApi.Dtos
