+++ DiscordBotApi/Handlers/BasicAuthenticationHandler.cs
@@ -23,8 +23,8 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-using DiscordBotApi.Data;
-using DiscordBotApi.Models;
+using DiscordBotApiLib.Data;
+using DiscordBotApiLib.Models;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
