+++ DiscordBotApiLib/IServeLogItemDiscordBotApi.cs
@@ -1,0 +1,14 @@
+﻿using DiscordBotApiLib.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DiscordBotApiLib
+{
+    public interface IServeLogItemDiscordBotApi
+    {
+        Task<ServerLogItem> DoSHit();
+    }
+}
